基本完成后台代码开发，文件上传解析功能完成，工作文件生成功能完成，数据库数据录入功能完成

+++ party_project/src/main/java/edu/gduf/domain/Mark.java
@@ -44,7 +44,7 @@ public class Mark implements Serializable {
     /**
      * 平均分排名
      */
-    private Double academicRanking;
+    private Integer academicRanking;
 
     public String getNum() {
         return num;
@@ -94,11 +94,11 @@ public class Mark implements Serializable {
         this.academicScore = academicScore;
     }
 
-    public Double getAcademicRanking() {
+    public Integer getAcademicRanking() {
         return academicRanking;
     }
 
-    public void setAcademicRanking(Double academicRanking) {
+    public void setAcademicRanking(Integer academicRanking) {
         this.academicRanking = academicRanking;
     }
